move some things around
