+++ crates/polars-lazy/src/lib.rs
@@ -192,6 +192,7 @@ pub mod dsl;
 pub mod frame;
 pub mod physical_plan;
 pub mod prelude;
+
 #[cfg(test)]
 mod tests;
 pub mod utils;
